# openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Wrtr REST Api
  description: >
    This is the document describing the API of Wrtr
  version: "0.6.0"
paths:
...
  /api/posts/:
    get:
      summary: Get posts made by user
      description: Get posts made by user, paginated
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            description: UUID of the user
        - name: page
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'schemas/post.yaml'
        "404":
          description: User was not found
  /api/posts/feed:
    get:
      summary: Get posts from user's feed
      description: Get posts from user's feed, paginated
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'schemas/post.yaml'
...